added rule for jsxExpression spacing '{ this.function() }'
