+++ test/jsxExpressionSpacingRule/tslint.json
@@ -1,0 +1,6 @@
+{
+    "rules": {
+        "jsx-expression-spacing": true
+    },
+    "rulesDirectory": "../../rules"
+}
